Re-design: - Removed excessive OrmObject class - removed equals method from PersistableObject - equals method of the removed OrmObject was moved to CompositeKey
